+++ project/backend/app/db/make_init_item_DB.py
@@ -1,6 +1,6 @@
 import os
 import psycopg
-from project.backend.app.utils.settings import load_backend_env
+from project.backend.app.manage.settings import load_backend_env
 
 # 1. 환경변수에서 Neon DB 접속 주소 불러오기
 load_backend_env()
